% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hic_export.R
\name{hic_export}
\alias{hic_export}
\title{Exports HiC file}
\usage{
hic_export(x, file)
}
\arguments{
\item{x}{A GAlignmentPairs object}

\item{file}{path or name of the output file}
}
\value{
A sequence file with the specified format
}
\description{
Exports an GAlignment object as bedpe file
}
\examples{
x <- hic_import("example.bam")
hic_export(x, "example.bedpe")
}
