% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hic_import.R
\name{hic_import}
\alias{hic_import}
\title{Imports HiC file}
\usage{
hic_import(file)
}
\arguments{
\item{file}{The path of the input file to read}
}
\value{
A GAlignmentPairs object
}
\description{
Imports HiC file
}
\examples{
x <- hic_import("example.bam")
}
